Bunch of files that I haven't commited.

+++ MyFinances/Controllers/BillsController.cs
@@ -9,6 +9,8 @@ using System.Web;
 using System.Web.Mvc;
 using MyFinances.Models;
 using Microsoft.AspNet.Identity;
+using MyFinances;
+using Microsoft.AspNet.Identity.Owin;
 
 namespace Finances.Controllers
 {
@@ -16,37 +18,46 @@ namespace Finances.Controllers
     public class BillsController : Controller
     {
         private ApplicationDbContext db = new ApplicationDbContext();
-        
-        public ActionResult Index(bool showInactive = false)
+        private ApplicationUserManager userManager;
+        private ApplicationUser user;
+
+        private void LoadUser ()
+        {
+            userManager = HttpContext.GetOwinContext().GetUserManager<ApplicationUserManager>();
+            user = userManager.FindById(User.Identity.GetUserId());
+        }
+
+        public ActionResult Index (bool showInactive = false)
         {
             ViewBag.ShowInactive = showInactive;
-            string userId = User.Identity.GetUserId();
+            LoadUser();
             if (showInactive)
             {
-                return View(db.Bills.Where(x => x.UserId == userId).Include(x => x.BillPayments).OrderBy(x => x.IsActive).ThenBy(x => x.DueDate).ToList().Populate());
+                return View(db.Bills.Where(x => x.UserId == user.Id).Include(x => x.BillPayments).OrderBy(x => x.IsActive).ThenBy(x => x.DueDate).ToList().Populate());
             }
             else
             {
-                return View(db.Bills.Where(x => x.UserId == userId && x.IsActive).Include(x => x.BillPayments).OrderBy(x => x.DueDate).ToList().Populate());
+                return View(db.Bills.Where(x => x.UserId == user.Id && x.IsActive).Include(x => x.BillPayments).OrderBy(x => x.DueDate).ToList().Populate());
             }
         }
-        
-        public ActionResult Details(int? id)
+
+        public ActionResult Details (int? id)
         {
             ViewBag.Action = "Details";
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
+            LoadUser();
             Bill bill = db.Bills.Find(id).Populate();
-            if (bill == null)
+            if (bill == null || bill.UserId != user.Id)
             {
                 return HttpNotFound();
             }
             return View(bill);
         }
-        
-        public ActionResult Create()
+
+        public ActionResult Create ()
         {
             ViewBag.Action = "Create";
             return View("Details");
@@ -54,7 +65,7 @@ namespace Finances.Controllers
 
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "Name,DueDate,Amount,Payee,IsDueDateStaysSame,IsAmountStaysSame,IsShared")] Bill bill)
+        public ActionResult Create ([Bind(Include = "Name,DueDate,Amount,Payee,IsDueDateStaysSame,IsAmountStaysSame,IsShared,PaymentFrequency")] Bill bill)
         {
             ViewBag.Action = "Create";
             if (ModelState.IsValid)
@@ -66,15 +77,16 @@ namespace Finances.Controllers
             return View("Details", bill);
         }
 
-        public ActionResult Edit(int? id)
+        public ActionResult Edit (int? id)
         {
             ViewBag.Action = "Edit";
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
+            LoadUser();
             Bill bill = db.Bills.Find(id).Populate();
-            if (bill == null)
+            if (bill == null || bill.UserId != user.Id)
             {
                 return HttpNotFound();
             }
@@ -83,10 +95,16 @@ namespace Finances.Controllers
 
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit ([Bind(Include = "ID,Name,DueDate,Amount,Payee,IsDueDateStaysSame,IsAmountStaysSame,IsShared")] Bill subBill)
+        public ActionResult Edit ([Bind(Include = "ID,Name,DueDate,Amount,Payee,IsDueDateStaysSame,IsAmountStaysSame,IsShared,PaymentFrequency")] Bill subBill)
         {
             ViewBag.Action = "Edit";
-            Bill bill = db.Bills.Find(subBill.ID).MapSubmit(subBill).Populate();
+            LoadUser();
+            Bill bill = db.Bills.Find(subBill.ID);
+            if (bill == null || bill.UserId != user.Id)
+            {
+                return HttpNotFound();
+            }
+            bill = bill.MapSubmit(subBill).Populate();
             if (ModelState.IsValid)
             {
                 db.Entry(bill).State = EntityState.Modified;
@@ -96,27 +114,54 @@ namespace Finances.Controllers
             return View("Details", bill);
         }
 
-        public ActionResult Deactivate(int id)
+        public ActionResult Deactivate(int? id)
         {
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            LoadUser();
             Bill bill = db.Bills.Find(id);
+            if (bill == null || bill.UserId != user.Id)
+            {
+                return HttpNotFound();
+            }
             bill.IsActive = false;
             db.Entry(bill).State = EntityState.Modified;
             db.SaveChanges();
             return RedirectToAction("Index");
         }
 
-        public ActionResult Activate (int id)
+        public ActionResult Activate (int? id)
         {
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            LoadUser();
             Bill bill = db.Bills.Find(id);
+            if (bill == null || bill.UserId != user.Id)
+            {
+                return HttpNotFound();
+            }
             bill.IsActive = true;
             db.Entry(bill).State = EntityState.Modified;
             db.SaveChanges();
             return RedirectToAction("Details", new { id = bill.ID });
         }
         
-        public ActionResult Delete(int id)
+        public ActionResult Delete(int? id)
         {
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            LoadUser();
             Bill bill = db.Bills.Find(id).Populate();
+            if (bill == null || bill.UserId != user.Id)
+            {
+                return HttpNotFound();
+            }
             if (bill.BillPayments.Any())
             {
                 bill.BillPayments.ToList().ForEach(x => db.BillPayments.Remove(x));
@@ -126,10 +171,19 @@ namespace Finances.Controllers
             return RedirectToAction("Index");
         }
 
-        public ActionResult AddPayment (int id)
+        public ActionResult AddPayment (int? id)
         {
             ViewBag.Action = "Add Payment";
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            LoadUser();
             Bill bill = db.Bills.Find(id).Populate();
+            if (bill == null || bill.UserId != user.Id)
+            {
+                return HttpNotFound();
+            }
 
             BillPayment payment = new BillPayment();
             payment.Amount = bill.Amount;
@@ -142,26 +196,44 @@ namespace Finances.Controllers
 
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult AddPayment (int id, [Bind(Include = "Amount,DatePaid,Payee")] BillPayment billPayment)
+        public ActionResult AddPayment (int? id, [Bind(Include = "Amount,DatePaid,Payee")] BillPayment billPayment)
         {
             ViewBag.Action = "Add Payment";
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            LoadUser();
             billPayment.Bill = db.Bills.Find(id).Populate();
+            if (billPayment == null || billPayment.Bill.UserId != user.Id)
+            {
+                return HttpNotFound();
+            }
             if (ModelState.IsValid)
             {
                 db.BillPayments.Add(billPayment);
-                db.Entry(saveBillForNextMonth(billPayment)).State = EntityState.Modified;
+                db.Entry(saveBillForNextPayment(billPayment)).State = EntityState.Modified;
                 db.SaveChanges();
                 return RedirectToAction("Details", new { id = billPayment.Bill.ID });
             }
             return View("Payment", billPayment);
         }
 
-        public ActionResult EditPayment (int id)
+        public ActionResult EditPayment (int? id)
         {
             ViewBag.Action = "Edit Payment";
-            BillPayment payment = db.BillPayments.Find(id);
-            payment.Bill = payment.Bill.Populate();
-            return View("Payment", payment);
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            LoadUser();
+            BillPayment billPayment = db.BillPayments.Find(id);
+            if (billPayment == null || billPayment.UserId != user.Id)
+            {
+                return HttpNotFound();
+            }
+            billPayment.Bill = billPayment.Bill.Populate();
+            return View("Payment", billPayment);
         }
 
         [HttpPost]
@@ -169,7 +241,12 @@ namespace Finances.Controllers
         public ActionResult EditPayment (int id, [Bind(Include = "Amount,DatePaid,Payee")] BillPayment billPayment)
         {
             ViewBag.Action = "Edit Payment";
+            LoadUser();
             BillPayment bp = db.BillPayments.Find(id);
+            if (bp == null || bp.UserId != user.Id)
+            {
+                return HttpNotFound();
+            }
             if (ModelState.IsValid)
             {
                 bp.Amount = billPayment.Amount;
@@ -182,18 +259,36 @@ namespace Finances.Controllers
             return View("Payment", billPayment);
         }
 
-        public ActionResult DeletePayment (int id)
+        public ActionResult DeletePayment (int? id)
         {
-            BillPayment bp = db.BillPayments.Find(id);
-            Bill bill = bp.Bill;
-            db.BillPayments.Remove(bp);
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            LoadUser();
+            BillPayment billPayment = db.BillPayments.Find(id);
+            if (billPayment == null || billPayment.UserId != user.Id)
+            {
+                return HttpNotFound();
+            }
+            Bill bill = billPayment.Bill;
+            db.BillPayments.Remove(billPayment);
             db.SaveChanges();
             return RedirectToAction("Details", new { id = bill.ID });
         }
 
-        public ActionResult Paid (int id, string r = "Details")
+        public ActionResult Paid (int? id, string r = "Details")
         {
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            LoadUser();
             Bill bill = db.Bills.Find(id);
+            if (bill == null || bill.UserId != user.Id)
+            {
+                return HttpNotFound();
+            }
             BillPayment billPayment = new BillPayment();
             billPayment.Bill = bill;
 
@@ -202,9 +297,16 @@ namespace Finances.Controllers
             billPayment.Payee = bill.Payee;
 
             db.BillPayments.Add(billPayment);
-            db.Entry(saveBillForNextMonth(billPayment)).State = EntityState.Modified;
+            db.Entry(saveBillForNextPayment(billPayment)).State = EntityState.Modified;
             db.SaveChanges();
-            return RedirectToAction(r, new { id = bill.ID });
+            if (r == "Dashboard")
+            {
+                return RedirectToAction("Index", "Dashboard");
+            }
+            else
+            {
+                return RedirectToAction(r, new { id = bill.ID });
+            }
         }
 
         protected override void Dispose(bool disposing)
@@ -216,26 +318,17 @@ namespace Finances.Controllers
             base.Dispose(disposing);
         }
 
-        private Bill saveBillForNextMonth(BillPayment billPayment)
+        private Bill saveBillForNextPayment(BillPayment billPayment)
         {
             Bill bill = billPayment.Bill;
-            DateTime nextDate = billPayment.DatePaid.AddMonths(1);
-            if (nextDate > bill.LastPaymentDate)
-            {
-                BillAverage avg = bill.BillAverages.Where(x => x.Date.Month == nextDate.Month).FirstOrDefault();
-                if (bill.IsDueDateStaysSame || avg == null)
-                {
-                    bill.DueDate = nextDate;
-                }
-                else
-                {
-                    bill.DueDate = avg.Date;
-                }
-                if (!bill.IsAmountStaysSame && avg != null)
-                {
-                    bill.Amount = avg.Average;
-                }
-            }
+
+            Tuple<DateTime, decimal> newDateAmount = bill.GetDateAmount(
+                bill.GetNextDate(billPayment.DatePaid, user),
+                bill.Amount);
+
+            bill.DueDate = newDateAmount.Item1;
+            bill.Amount = newDateAmount.Item2;
+
             return bill;
         }
     }
